Fix dockerng.network_* name matching
Updating the states so that they loop through the list of networks
returned from `dockerng.networks` to check that one of the networks
returned fully matches the name of the network that needs to be ensured
to be present or absent.

Another option would be to add an argument to the `dockerng.networks`
execution module to enable strict matching.  This would probably be
preferable to repeating the logic in the states, but I didn't want to
meddle too much as the execution module is currently just wrapping the
docker-py library.

Fixes #41982